Move entities to a separate module
This makes it easier to grep through the known elements.
