+++ html5lib/_tokenizer.py
@@ -2,11 +2,10 @@ from collections import deque, OrderedDict
 from sys import version_info
 
 from .constants import spaceCharacters
-from .constants import entities
 from .constants import asciiLetters, asciiUpper2Lower
 from .constants import digits, hexDigits, EOF
 from .constants import tokenTypes, tagTokenTypes
-from .constants import replacementCharacters
+from .entities import entities, replacementCharacters
 
 from ._inputstream import HTMLInputStream
